+++ src/abstractions/browser/commands/pasteImage.ts
@@ -1,6 +1,5 @@
 import * as vscode from "vscode";
-import { createCommand } from "../../../utils/createCommand";
 
-export const pasteImageCommand = createCommand(() => {
+export async function pasteImageCommand() {
   vscode.window.showErrorMessage("Paste image is not supported in the browser");
-});
+}