+++ src/interfaces/IClipboardToImageBuffer.ts
@@ -1,5 +1,0 @@
-
-// abstraction over the clipboard
-export interface IClipboardToImageBuffer {
-    getImageBits(): Promise<Buffer>;
-}