+++ src/interfaces/IPlaygroundJSON.ts
@@ -1,6 +1,0 @@
-export interface IPlaygroundJSON {
-  scripts?: string[];
-  styles?: string[];
-  layout?: string;
-  showConsole?: boolean;
-}